+++ framework/server/src/main/java/is/codion/framework/server/EntityServer.java
@@ -532,7 +532,7 @@ public class EntityServer extends AbstractServer<AbstractRemoteEntityConnection,
 
 		@Override
 		public void run() {
-			database.closeConnectionPools();
+			database.close();
 		}
 	}